Cast int to double to avoid arithmetic overflow warning

+++ include/wx/math.h
@@ -136,7 +136,7 @@ inline bool wxIsNullDouble(double x) { return wxIsSameDouble(x, 0.); }
 
 inline int wxRound(double x)
 {
-    wxASSERT_MSG( x > INT_MIN - 0.5 && x < INT_MAX + 0.5,
+    wxASSERT_MSG( x > (double)INT_MIN - 0.5 && x < (double)INT_MAX + 0.5,
                   wxT("argument out of supported range") );
 
     #if defined(HAVE_ROUND)
